libstdc++: Fix non-portable results from 64-bit std::subtract_with_carry_engine [PR107466]
This fixes a regression introduced by the LWG 3809 change, so is needed
on trunk and gcc-13 and gcc-12.

Tested x86_64-linux and aarch64-linux.

-- >8 --

I implemented the resolution of LWG 3809 in r13-4364-ga64775a0edd469 but
more recently it was noted that the change causes possible truncation
for 64-bit seeds. Whether the truncation occurs (and to what value)
depends on the width of uint_least32_t which is not portable. The new
problem was filed as LWG 4014. I proposed a new change which reduces the
seed by the LCG's modulus before the conversion to uint_least32_t. This
ensures a consistent result across paltforms, and restores the old
behaviour for std::subtract_with_carry_engine specializations using a
64-bit result type (such as std::ranlux48_base).

libstdc++-v3/ChangeLog:

	PR libstdc++/107466
	* include/bits/random.tcc (subtract_with_carry_engine::seed):
	Implement proposed resolution of LWG 4014.
	* testsuite/26_numerics/random/pr60037-neg.cc: Adjust dg-error
	line number.
	* testsuite/26_numerics/random/subtract_with_carry_engine/cons/lwg3809.cc:
	Check for expected result of 64-bit engine with seed that
	doesn't fit in 32-bits.