Fix PR107201 -nodevicelib not working for all devices.
Since the advent of devices AVR*, the spec pattern mmcu=avr* does no
more work to discriminate between devices and cores like avr51.

This means -nodevicelib no more works for AVR* devices because that
option is removed for mmcu=avr* (which were only cores in the old days).

Instead of that pattern, a new spec function is used.

When there are no objections or improvements within the next week or
so, I'd install the patch below.

Johann