libstdc++: Update tzdata to 2023d
On Fri, 12 Jan 2024 at 22:59, Jonathan Wakely wrote:
>
> It would be good to update the bundled tzdata for GCC 14.1 and 13.3

The expiry date for the hardcoded leapseconds list should be updated
too, as there's a new date in the file in the tzdata distro. There are
no new leap seconds though, just a new "this list is valid until ..."
date.

Tested x86_64-linux and aarch64-linux.
commit e65bd5d069ad27ce01854ad52224bc0186c924cf
Author: Jonathan Wakely <jwakely@redhat.com>
Date:   Fri Jan 12 16:57:41 2024

    libstdc++: Update tzdata to 2023d

    Import the new 2023d tzdata.zi file. The leapseconds file was also
    updated to have a new expiry (no new leap seconds were added).

    libstdc++-v3/ChangeLog:

            * src/c++20/tzdata.zi: Import new file from 2023d release.
            * src/c++20/tzdb.cc (tzdb_list::_Node::_S_read_leap_seconds)
            Update expiry date for leap seconds list.