libstdc++/ranges: Define _S_has_simple_call_op on newer adaptors
Tested on x86_64-pc-linux-gnu, does this look OK for trunk?

-- >8 --

This compile-time and diagnostic improvement[1] is less important in
C++23 mode where deducing this is available and used in _Pipe, but for
benefit of C++20 mode and for consistency let's set the flag on the most
recently added range adaptors which lack it.

[1]: Defining _S_has_simple_call_op=true for a range adaptor closure
object signals that the adaptor has an operator() that's not overloaded
according to the constness/value category of the object, which in turn
allows us to implement the operator() of the composition of such adaptors
in a simpler way.

libstdc++-v3/ChangeLog:

	* include/std/ranges (views::_Adjacent::_S_has_simple_call_op):
	Define to true.
	(views::_AsRvalue::_S_has_simple_call_op): Likewise.
	(views::_Enumerate::_S_has_simple_call_op): Likewise.
	(views::_AsConst::_S_has_simple_call_op): Likewise.