xfail libgomp.c/declare-variant-4-{fiji,gfx803}.c
The problem is as described at
https://gcc.gnu.org/install/specific.html#amdgcn-x-amdhsa

"Note that support for Fiji devices has been removed in ROCm 4.0 and
support in LLVM is deprecated and will be removed in LLVM 18."

Therefore, GCC is no longer build with Fiji (gfx803) support by default
– and the -march=fiji testcases now fails as the -lgomp multilib for
Fiji is not available. (That is: It fails, unless Fiji support has been
enabled manually.)

Andrew mentioned that there is a PR about this, but I couldn't find it.
If someone can, I am happy to add it to the changelog.

OK for mainline?

Tobias

xfail libgomp.c/declare-variant-4-{fiji,gfx803}.c

Since r14-4734-g56ed1055b2f40ac162ae8d382280ac07a33f789f, GCC no longer
builds the Fiji (alias gfx803) libraries by default as support for it was
removed in ROCm 4.0 and will be removed in LLVM 18.

Thus, unless gfx803 is explicitly enabled, the following testcases will
fail to link as libgomp is not available for Fiji. Hence, this commit
xfails those testcases.

libgomp/ChangeLog:

	* testsuite/libgomp.c/declare-variant-4-fiji.c: Xfail as fiji
	support is no longer enabled by default.
	* testsuite/libgomp.c/declare-variant-4-gfx803.c: Likewise.

Signed-off-by: Tobias Burnus <tburnus@baylibre.com>

 libgomp/testsuite/libgomp.c/declare-variant-4-fiji.c   | 2 ++
 libgomp/testsuite/libgomp.c/declare-variant-4-gfx803.c | 2 ++
 2 files changed, 4 insertions(+)